requirements.txt
numpy==1.24.3
pinecone-client==2.2.1
matplotlib==3.7.1
